+++ src/orcaset/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2026 Orcaset Inc.
 # SPDX-License-Identifier: SSPL-1.0
 
-from orcaset.context import Context, CycleError, DepNode
+from orcaset.context import Context, ConvergenceError, CycleError, DepNode
 from orcaset.formatters import (
     DateFormatter,
     ValueFormatter,
@@ -12,7 +12,17 @@ from orcaset.formatters import (
 from orcaset.maybe import Maybe, Na, add_values, combine_values, isna, map2_some, map_some
 from orcaset.period import Period, date_union, period_union
 from orcaset.query import DayCount, accrual, accrual_or, exact, exact_or, last
-from orcaset.rule import Demand, KeyedRule, Rule, Step, get, get_at
+from orcaset.rule import (
+    Demand,
+    Iterate,
+    KeyedRule,
+    Rule,
+    Step,
+    abs_distance,
+    get,
+    get_at,
+    maybe_abs_distance,
+)
 from orcaset.series import (
     BaseSeries,
     CellFactory,
@@ -52,6 +62,7 @@ __all__ = [
     "CellStream",
     "CellsFn",
     "Context",
+    "ConvergenceError",
     "CycleError",
     "DateFormatter",
     "DateSeries",
@@ -62,6 +73,7 @@ __all__ = [
     "DepNode",
     "Group",
     "GroupRow",
+    "Iterate",
     "Key",
     "KeyedRule",
     "LineRow",
@@ -86,6 +98,7 @@ __all__ = [
     "Total",
     "TotalRow",
     "ValueFormatter",
+    "abs_distance",
     "accrual",
     "accrual_or",
     "add_values",
@@ -102,5 +115,6 @@ __all__ = [
     "map2_some",
     "map_some",
     "markdown_table",
+    "maybe_abs_distance",
     "period_union",
 ]